requirements.txt
# 
# requirements.txt
# Used for local development.  May resolve to GPU-capable builds if CUDA is
# present on the host.  The Dockerfile pins to CPU-only PyTorch explicitly.
# 

torch>=2.1.0
torchvision>=0.16.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
